Moved testing of stopwatch into testing package
