+++ src/net/sf/memoranda/test/MemorandaTestSuite.java
@@ -6,6 +6,6 @@ import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-        EventImplTest.class, DefectListCreationTest.class, DefectListPullingFromFileTest.class
+        EventImplTest.class, DefectListCreationTest.class, DefectListPullingFromFileTest.class, StopwatchTest.class
 })
 public class MemorandaTestSuite {}
